+++ ICD.Connect.Displays.Panasonic/Devices/PanasonicClassicDisplay.cs
@@ -205,7 +205,9 @@ namespace ICD.Connect.Displays.Panasonic.Devices
 		{
 			if (!VolumeControlAvailable)
 				return;
-			string setVolCommand = GenerateSetVolumeCommand((int)raw);
+
+			int volume = (int)Math.Round(raw);
+			string setVolCommand = GenerateSetVolumeCommand(volume);
 			SendNonFormattedCommand(setVolCommand);
 			SendNonFormattedCommand(QUERY_VOLUME);
 		}